(PT-426) Install app in arbitrary cluster
This pulls in work Michael did in puppetlabs/kurl_test#19 to allow
installation in clusters other than kurl. In particular it provides for
installing into GKE. It will start an ingress and determine the ingress
ip, and if your target is localhost and no hostname is provided directly
or indirectly from provided config, it will construct a nip.io address
for the app from the load balancer ip. This allows the wait_for_app task
to connect and complete it's readiness tests.

A separate task to return load balancer ip was added, and the code to
retun the ip when starting ingress was removed, since we may be running
against a cluster with an lb already up.