(REPLATS-426) Update get_kots_app_status to check kotsadm state
If kotsadm is not installed, the task can't interact with it via
`kubectl-kots get app` to determine individual app status. However, this
isn't a straight up error state, because `kubectl-kots install` will
first install kots itself if not present. So the task now looks to see
if it can find a kotsadm pod, and if it cannot, it treats this as the same
app 'not-installed' state so that install plans can proceed.