(REPLATS-427) Add function to generate a random name
Will be using this in kurl_test to generate names for randomized working
directories. Having this as a Puppet function means we can use a single
tempdir for a plan's array of targets and not have to match, for example,
`mktemp -d` output across targets. It's pretty generic, so placing it in
pam_tools in case we find uses there as well.