(REPLATS-457) Add some acceptance testing of tasks and plans
...to the github actions. Installs bolt and kubectl tools into a k3d
environment and runs the plans and tasks against localhost for some
basic acceptance testing.

Tests:

* install_published/teardown plans (which also exercise the underlying
  kots_install, wait_for_app and delete* tasks)
* helm install/delete tasks

Also has the side benefit of validating that we run in k3d/k3s.

Could certainly do more here to validate the state post task/plan
execution, but at least this is ensuring that the main tasks and plans
run.