(maint) Remove default --skip-preflights from kots_install_options
This was set in kurl_test, because the kurl vmpooler image is smaller
than preflights will accept for Connect. But bypassing preflights is
really something the caller should decide about, so I'm removing it from
pam_tools.

This is a breaking change that will require a fix in kurl_test to pass
it back in the kurl_test::install_replicated_app plan.