(REPLATS-457) Make pam channel settable for kots_install
The kots_install task had puppet-application-manager/stable hardset as
the slug/channel used for initial configuration of Kots. This becomes an
issue in GKE clusters with more restrictive RBAC since Kots expects to
be able to manipulate clusterroles when first installed. By making this
configurable, you can set +pam_variant+ to minimal-rbac and `kots
install` will be called with 'puppet-application-manager/minimal-rbac'
instead. This could be used to test other pam channels as well...