(maint) Bump json version for CVE-2020-10663
Snyk noted that there was a JSON 2.1.0 vulnerability:
https://app.snyk.io/vuln/SNYK-RUBY-JSON-560838

The default pdk Gemfile had some very specific json version pins based
on ruby version, which I am arbitrarily throwing out in favor of
grabbing latest. If I get rid of the dependency altogether, you will
end up with the json in the ruby stdlib, which according to
https://stdgems.org/2.6.8/, puts you back with 2.1.0, on Ruby 2.6.8, for
example, which happens to be what I was running locally atm.

At any rate, unit tests, at least, pass with 2.5.z, so we'll run with
that until we find out why it doesn't work.