(REPLATS-457) Extract app config generation from kots_install task
The task will still provide a very rudimentary config sufficient to
complete a generic install if no configuration is given, but it no
longer sets the resource constraints that were hard coded for the kurl
vm's 4cpu/8gb environment.

Config construction is pulled up to a Puppet template and a function for
very basic PE memory tuning calculation. The memory settings for connect
can be grossly changed by providing a total memory value to the
install_published plan.

The config defaults to 16GB, so the Github actions workflow now
specifies allocated_memory_in_megabytes of 7 to match the Github
runner's 7GB of memory.

When kurl_test picks up this version, it will be able to supply the
target memory to install_published, and the PE chart values template
will be able to use the pam_tools::calculcate_pe_memory() instead of
duplicating the calculation there.

Because this commit adds a spec/classes spec to test the template and a
spec/fixtures/manifests class to invoke the template, I needed to change
the spec_helper to not invoke BoltSpec::Plans.init(), as this
interferes with rspec-puppet using the compiler for Puppet class
testing (or possibly with finding the right environment?).
Interestingly, removing this entirely didn't affect the running of the
existing plan specs. So something may have changed in recent Bolt
versions.