(maint) Fix wait bug in delete app resources task
If the app you were deleting had an in progress job, the
delete_k8s_app_resources task would fail because after successfully
scaling down deployments and statefulsets it would wait for all pods
associated with the label to terminate. But an in progress job that is
now never going to complete can end up timing this out because wait
blocks on the job's still running pod.

Removed the wait command from the scale_down() method and instead moved
the timeout to the scale command call.