(maint) Delete all three resource labels from Makefile
The holodeck makefile is deleting three resource labels
when cleaning up after a replicated app installation. This task was
originally onrly written with the main app.kubernetes.io/part-of
label. These leaves behind a couple of secrets and newer vault work.
Added in the new labels so that we clear everything out.