(REPLATS-457) Return loadbalancer ip when starting ingress
After starting the nginx ingress, poll for the loadbalancer ip and return
it in the task data so it can be used by the caller.

Useful so that you know the ip to test your application on after
installation, and for setting up hostname/certs when using dev work
arounds like nip.io addresses that embed the ip to short circuit lack of
dns for ephemeral test instances.

This commit also includes two brain dead fixes for the start/destroy
tasks in that they previously did not run themselves when executed.