(maint) Remove kubeconfig parameter from helm chart install task
Testing has shown that it's sufficient to set the KUBECONFIG env
variable if you have a non-standard kubernetes config file you need to
reference, so I'm removing the kubeconfig parameter as it was blocking
that workflow. This was for the case of a localhost testrunner operating on
a remote GKE cluster. I think the same would be try for a remote runner;
either you would ensure that the default ~/.kube/config is what you need
or have KUBECONFIG exported in the environment Bolt will be shelling
into.