(REPLATS-457) Handle ingress for KinD clusters
The process for testing and starting ingress in a KinD cluster is
slightly different. To handle this, we need to be able to check whether
the cluster is KinD or not. This test is a little peculiar since what
makes it KinD is that the env is running in a docker container, but the
test we're using is looking for a KinD specific pod via kubectl.

If we tried to use a more specific check relying on docker such as
existence of a container with a known KinD label:

`docker ps -f 'label=io.x-k8s.kind.cluster' --format '{{.ID}}'`,

the issue becomes that the test will only work for ssh targets, and the
localhost case with local kubeconfig pointing to the kind cluster fails.

Given knowledge of whether the cluster is KinD or not, the changes for
has_ingress_controller and start_nginx_ingress are minimal.

The has_ingress_controller task checks NodePort.port instead of
NodePort.nodePort for KinD clusters.

This is because the Nginx ingress controller's kind provider starts up
a NodePort service with nodePorts randomly set to the default
30000-32767 range, rather than the 80/443 we would normally test for.
On the host running the container, docker-proxy proxies 80/443 to the
container's ip (if configured to do so when the cluster was created).
And the final translation of an incoming packet from container_ip:80 to
random_nodePort seems to happen in the iptables nat table as a dnat
rule. I'm not certain where that magic gets generated -- possibly in the
kindnet service?

At any rate, to test for whether the ingress is installed, we need to
check the port parameter in a KinD cluster.

For start_nginx_controller, in a KinD cluster, it needs to use the
'kind' provider instead of cloud, by default.