(maint) Skip kotsadm delete on Kurl by default
Deleting the kotsadm application from a Kurl cluster removes the
kurl-proxy which is not reinstalled when kotsadm is installed by the
next `kots install` command. Instead the kurl installer has to be
re-run.

To avoid this, if we detect that a kurl-proxy is running in the cluster,
skip attempting to delete kotsadm unless a +force+ flag is true.