(maint) Drop pdk transitive dependency on pluginator
The puppet-module-posix-dev is a gem contrivance used by pdk to
associate a bunch of dev gem dependencies under one version handler.
It's added by the pdk when generating a new module so that tools needed
by various Rake tasks for managing the module are available. It contains
a transitive dependency on pluginator (associated with the
puppet-debugger) which causes Snyk heartburn because it's lgpl.

We don't use the puppet-debugger, so this patch yanks the pdk gems out so as
to sidestep this. It directly includes the gems needed to get unit_test to
run.

This undoubtadedly breaks other Rake tasks we aren't using that depend
on other gems that were previously included by puppet-module-posix-dev.
So in the future we might need to add additional gems if we were to use
other tasks.